Move callTreeFromProfile to test utils
